+++ src/main/java/uk/gov/hmcts/reform/prl/services/managedocuments/ManageDocumentsService.java
@@ -3,24 +3,16 @@ package uk.gov.hmcts.reform.prl.services.managedocuments;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
-import com.sendgrid.Method;
-import com.sendgrid.Request;
-import com.sendgrid.Response;
 import com.sendgrid.SendGrid;
-import com.sendgrid.helpers.mail.Mail;
-import com.sendgrid.helpers.mail.objects.Email;
-import com.sendgrid.helpers.mail.objects.Personalization;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections.CollectionUtils;
-import org.apache.commons.collections.MapUtils;
 import org.apache.commons.io.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
@@ -96,8 +88,6 @@ import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.SOLICITOR;
 import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.SOLICITOR_ROLE;
 import static uk.gov.hmcts.reform.prl.models.complextypes.QuarantineLegalDoc.quarantineCategoriesToRemove;
 import static uk.gov.hmcts.reform.prl.services.ManageOrderEmailService.THERE_IS_A_FAILURE_IN_SENDING_EMAIL_TO_SOLICITOR_ON_WITH_EXCEPTION;
-import static uk.gov.hmcts.reform.prl.services.SendgridService.MAIL_SEND;
-import static uk.gov.hmcts.reform.prl.services.SendgridService.NOTIFICATION_TO_PARTY_SENT_SUCCESSFULLY;
 import static uk.gov.hmcts.reform.prl.utils.ElementUtils.element;
 import static uk.gov.hmcts.reform.prl.utils.ElementUtils.nullSafeCollection;
 
@@ -121,6 +111,8 @@ public class ManageDocumentsService {
     private final SendgridEmailTemplatesConfig sendgridEmailTemplatesConfig;
 
     private final SendGrid sendGrid;
+
+    private final ManageDocumentEmailService manageDocumentEmailService;
 
     @Value("${send-grid.rpa.email.from}")
     private String fromEmail;
@@ -383,7 +375,7 @@ public class ManageDocumentsService {
                                       String emailAddress,
                                       SendgridEmailTemplateNames sendgridEmailTemplateName) {
         try {
-            sendEmailUsingTemplateWithAttachments(
+            manageDocumentEmailService.sendEmailUsingTemplateWithAttachments(
                 sendgridEmailTemplateName,
                 authorisation,
                 SendgridEmailConfig.builder()
@@ -398,41 +390,6 @@ public class ManageDocumentsService {
         }
     }
 
-    public void sendEmailUsingTemplateWithAttachments(SendgridEmailTemplateNames sendgridEmailTemplateNames,
-                                                      String authorization, SendgridEmailConfig sendgridEmailConfig) throws IOException {
-        Personalization personalization = new Personalization();
-        personalization.addTo(getEmail(sendgridEmailConfig.getToEmailAddress()));
-        Map<String, Object> dynamicFields = sendgridEmailConfig.getDynamicTemplateData();
-        if (MapUtils.isNotEmpty(dynamicFields)) {
-            dynamicFields.forEach(personalization::addDynamicTemplateData);
-        }
-        Mail mail = new Mail();
-        mail.setFrom(getEmail(fromEmail));
-        mail.addPersonalization(personalization);
-        mail.setTemplateId(getTemplateId(sendgridEmailTemplateNames, sendgridEmailConfig.getLanguagePreference()));
-        Request request = new Request();
-        try {
-            request.setMethod(Method.POST);
-            request.setEndpoint(MAIL_SEND);
-            request.setBody(mail.build());
-            Response response = sendGrid.api(request);
-            log.info("Sendgrid status code {}", response.getStatusCode());
-            if (!HttpStatus.valueOf(response.getStatusCode()).is2xxSuccessful()) {
-                log.info(NOTIFICATION_TO_PARTY_SENT_SUCCESSFULLY);
-            }
-        } catch (IOException ex) {
-            log.info("error is {}", ex.getMessage());
-            throw new IOException(ex.getMessage());
-        }
-    }
-
-    private Email getEmail(String toEmailAddress) {
-        return new Email(toEmailAddress);
-    }
-
-    private String getTemplateId(SendgridEmailTemplateNames templateName, LanguagePreference languagePreference) {
-        return sendgridEmailTemplatesConfig.getTemplates().get(languagePreference).get(templateName);
-    }
 
     private QuarantineLegalDoc convertQuarantineDocumentToRightCategoryDocument(QuarantineLegalDoc quarantineLegalDoc, UserDetails userDetails) {
         String loggedInUserType = DocumentUtils.getLoggedInUserType(userDetails);
